Allow create of osd on mounted devices
Addes logic to create and activate partitions on devices that have
mounted partitions.
* Build ceph-disk flags in function to re-use current logic from
osdize_dev in osdize_part
* Shared osd's implement all options/logic of the osdize_dev
 * Checks for in use
 * Handles the format osd option
 * Works with dmcrypt
 * Will not format/zap non-ceph partitions
 * Will not format/zap mounted ceph partitions

This includes two drive-by fixes for bugs which it didn't make since to
break out. Both changes are in code which was altered in this
patch and not closing them would require implementing the bug to avoid
closing it.
* Partitions are wiped before zapping
* Zap is done as an independent call to ceph-disk

All bugs require syncing to relevant charms to actually close
Partial-Bug: #1743278
Partial-Bug: #1743280
Partial-Bug: #1743282

Change-Id: Id3c8403dfbf710e054ca5e4d965fe6f9b8f9c72b